SinoFilter: Move non-opencl related functions to silx.image.tomography
